refactor(envs): rename _helpers.py to utils.py and extract GroundingConfig
- src/oumi/environments/_helpers.py -> src/oumi/environments/utils.py:
  module is part of the public surface (consumed by GroundingConfig.describe_grounding
  paths), so the underscore prefix was misleading.
- src/oumi/core/configs/params/grounding_params.py: new home for GroundingConfig
  (extracted from environment_params.py to keep that file focused on environment
  params and avoid co-locating unrelated config types).
- utils.py now imports DeterministicToolOutput from the deterministic_tool sibling
  module rather than tool_params (Phase 1 moved the class).
- environments/__init__.py and the grounding tests updated to point at the new
  module paths.
