+++ src/oumi/core/configs/params/grounding_params.py
@@ -1,0 +1,46 @@
+# Copyright 2025 - Oumi
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Per-environment grounding configuration."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+from oumi.core.configs.params.base_params import BaseParams
+
+
+@dataclass
+class GroundingConfig(BaseParams):
+    """Per-environment grounding configuration.
+
+    When set on an environment, the ConversationSynthesizer samples facts from
+    that environment and injects them into the planner prompt so turn plans
+    reference real entities rather than hallucinated IDs.
+    """
+
+    sample_size: int = 3
+    """Number of grounding facts sampled per conversation."""
+
+    seed: int | None = None
+    """If set, per-sample RNG is seeded from ``(seed + sample_index)`` for
+    reproducibility. If None, grounding uses an unseeded ``random.Random``."""
+
+    def __post_init__(self) -> None:
+        """Validate ``sample_size`` is positive."""
+        if self.sample_size < 1:
+            raise ValueError(
+                f"{type(self).__name__}.sample_size must be >= 1, "
+                f"got {self.sample_size}."
+            )
