+++ src/oumi/environments/__init__.py
@@ -18,7 +18,7 @@ Importing this package populates the environment registry by triggering each
 concrete environment's `@register_environment(...)` decorator.
 """
 
-from oumi.core.configs.params.environment_params import GroundingConfig
+from oumi.core.configs.params.grounding_params import GroundingConfig
 from oumi.core.configs.params.tool_params import (
     ToolArgumentError,
     ToolError,
